# openapi/v1/openapi.yaml
openapi: 3.1.0
info:
  title: ScamNemesis API
  description: |
    REST API for the ScamNemesis fraud reporting and tracking platform.

    ## Authentication
    The API supports three authentication methods:
    - **Bearer Token**: JWT access token in Authorization header
    - **HttpOnly Cookie**: Automatic for browser clients
    - **API Key**: X-API-Key header for programmatic access

    ## Rate Limiting
    - Standard users: 100 requests/hour per endpoint
    - Admin users: 1000 requests/hour (10x multiplier)
    - Auth endpoints: 20 requests/15min (brute force protection)

    ## Error Handling
    All errors follow RFC 7807 Problem Details format with additional fields.
  version: 1.0.0
  contact:
    name: ScamNemesis Support
    url: https://scamnemesis.com/contact
    email: support@scamnemesis.com
  license:
    name: Proprietary
    url: https://scamnemesis.com/terms
servers:
  - url: https://scamnemesis.com/api/v1
    description: Production server
  - url: https://staging.scamnemesis.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - BearerAuth: []
  - CookieAuth: []
  - ApiKeyAuth: []
paths:
...
  /evidence/upload:
    post:
      tags: [Evidence]
      summary: Upload evidence file
      description: Upload evidence file (max 10MB, images/documents)
      operationId: uploadEvidence
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [SCREENSHOT, DOCUMENT, CONVERSATION, TRANSACTION, OTHER]
                description:
                  type: string
                  maxLength: 500
              required: [file, type]
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_key:
                    type: string
                  url:
                    type: string
        '400':
          description: Invalid file
        '413':
          description: File too large
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /contact:
    post:
      tags: [Contact]
      summary: Submit contact form
      description: Submit a contact form message
      operationId: submitContact
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                email:
                  type: string
                  format: email
                subject:
                  type: string
                  maxLength: 200
                message:
                  type: string
                  maxLength: 5000
                captchaToken:
                  type: string
              required: [name, email, subject, message]
      responses:
        '201':
          description: Message sent
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ============ CAPTCHA ============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly cookie authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
    WebhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Webhook secret for external services
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        request_id:
          type: string
          description: Request ID for debugging
        timestamp:
          type: string
          format: date-time
      required: [error, message]
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            details:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
            issues:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
  responses:
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: rate_limited
            message: Rate limit exceeded. Try again later.